Add more statistics tracked to character saves #578 (#589)
* Add extensions for working with statistics

* Use statistic extension in BreakEventXPListener.cs

* Add SkillExtensions.cs

* Add break and place listeners

* Fix stats extensions not adding missing stats

* Fix statistics not saving

* change statistics to be a long

* Use statistics for last played and age

* cleanup

* revert using stats for age and last played

* rename listeners

* add traveled statistic

* Add travel stats

* Version up character data

+++ WaywardBeyond.Client.Core/Events/XPEvent.cs
@@ -2,10 +2,10 @@ using WaywardBeyond.Client.Core.Skills;
 
 namespace WaywardBeyond.Client.Core.Events;
 
-internal readonly struct XPEvent(Skill skill, int level, int xp, int gainedXP)
+internal readonly struct XPEvent(Skill skill, int level, long xp, long gainedXP)
 {
     public readonly Skill Skill = skill;
     public readonly int Level = level;
-    public readonly int XP = xp;
-    public readonly int GainedXP = gainedXP;
+    public readonly long XP = xp;
+    public readonly long GainedXP = gainedXP;
 }
